Updated Services, added nswag spescs (..)
- implemented clientside phases.
- wrote a simple background.
- fixed few errors
- plugged most of the features.
- fixed errors were wasn't possible to save probes.
- added order to probes.
- added actions.

+++ Core/Models/Grid.cs
@@ -6,7 +6,7 @@ namespace Core.Models
 {
 	public class Grid
 	{
-		public byte X { get; set; }
-		public byte Y { get; set; }
+		public int X { get; set; }
+		public int Y { get; set; }
 	}
 }
